Extract Trainjob from the Kubeflow directory
